+++ dev-packages/node-integration-tests/suites/tracing/httpIntegration/instrument-options.mjs
@@ -19,31 +19,6 @@ Sentry.init({
           resMethod: res.req.method,
         });
       },
-      instrumentation: {
-        requestHook: (span, req) => {
-          span.setAttribute('attr1', 'yes');
-          Sentry.setExtra('requestHookCalled', {
-            url: req.url,
-            method: req.method,
-          });
-        },
-        responseHook: (span, res) => {
-          span.setAttribute('attr2', 'yes');
-          Sentry.setExtra('responseHookCalled', {
-            url: res.req.url,
-            method: res.req.method,
-          });
-        },
-        applyCustomAttributesOnSpan: (span, req, res) => {
-          span.setAttribute('attr3', 'yes');
-          Sentry.setExtra('applyCustomAttributesOnSpanCalled', {
-            reqUrl: req.url,
-            reqMethod: req.method,
-            resUrl: res.req.url,
-            resMethod: res.req.method,
-          });
-        },
-      },
     }),
   ],
 });